requirements.txt
# RAG Query Test Dependencies
python-dotenv>=1.1.1
pinecone-client>=3.0.0
openai>=1.12.0
google-generativeai>=0.8.5

# Local development server
flask>=3.0.0
flask-cors>=4.0.0

# Ingestion Dependencies
GitPython>=3.1.30
boto3>=1.26.0
llama-index-core>=0.12.0
PyYAML>=6.0

# Sparse Search
typesense>=0.25.0
opensearch-py>=2.0.0

# Core dependencies (automatically installed with above)
certifi>=2025.8.3
typing-extensions>=4.14.1
urllib3>=2.5.0
requests>=2.32.4
pydantic>=2.11.7
anyio>=4.10.0